Add shared file model records
