--- Filters/NotFoundExceptionFilter.cs
@@ -1,17 +1,0 @@
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-using System.IO;
-
-namespace TestProject.Filters;
-
-public class NotFoundExceptionFilter : IExceptionFilter
-{
-    public void OnException(ExceptionContext context)
-    {
-        if (context.Exception is FileNotFoundException or DirectoryNotFoundException)
-        {
-            context.Result = new NotFoundResult();
-            context.ExceptionHandled = true;
-        }
-    }
-}
